requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
cloudinary==1.20.0
Django==3.0.6
django-cloudinary-storage==0.2.3
django-debug-toolbar==2.2
django-seed==0.2.2
djangorestframework==3.11.0
Faker==4.1.0
idna==2.9
Pillow==7.1.2
psycopg2-binary==2.8.5
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.9
